KAFKA-5142: Corrected checkstyle and refactored SimpleHeaderConverter
